fix: inference pipeline (#6)
* feat: first inference pipeline

* feat: minus mask on whole image

* feat: first inference from sentinel

* feat: inference first pipeline

* docs: images

* feat: onion base

* fix: tests

* feat: inference is working!

* feat: inference with sentinel

* style: typing

* refactor: delete old files

* refactor: move files

* feat: previz and color normalization

* perf: opti

* feat: changes

* refactor: download bands

* refactor: onionize

* refactor: download and preview

* docs: update documentation and pre-commit

* bkp

* it's fixed!

+++ satellite/src/infrastructure/model.py
@@ -73,4 +73,4 @@ class TorchModelService(ModelService):
         return model
 
     def predict(self, tile: Tile) -> np.ndarray:
-        return self.model(torch.from_numpy(tile.data).permute(2, 0, 1).unsqueeze(0)).squeeze().detach().numpy()
+        return self.model(torch.from_numpy(tile.data).permute(2, 0, 1).unsqueeze(0).float()).squeeze().detach().numpy()
